bugfix: S3C-3114 close inactive free sockets in http agent
Automatically close free sockets in the HTTP/HTTPS agent that have
been inactive for a certain time. The inactivity time is set to 55
seconds, which is lower than the configured 60 seconds timeout for
idle connections on bucketd server, avoiding the risk of reusing a
connection that is just being closed by the server, resulting in an
ECONNRESET error.

It uses the third-party npm module agentkeepalive which provides the
extra "freeSocketTimeout" agent option, not available in the regular
node.js http.Agent.